update of handlers
