+++ softgroup/services/authorization/authorization-impl/src/main/java/handler/AuthorizationRequestLoginHandler.java
@@ -1,0 +1,31 @@
+package handler;
+
+import com.softgroup.authorization.api.message.LoginRequest;
+import com.softgroup.authorization.api.message.LoginResponse;
+import com.softgroup.authorization.api.router.AuthorizationRequestHandler;
+import com.softgroup.authorization.api.router.AuthorizationRouterHandler;
+import com.softgroup.common.protocol.Request;
+import com.softgroup.common.protocol.Response;
+import com.softgroup.common.router.api.AbstractRequestHandler;
+import org.springframework.stereotype.Component;
+
+/**
+ * Created by polev on 25.02.2017.
+ */
+
+@Component
+public class AuthorizationRequestLoginHandler
+        extends AbstractRequestHandler<LoginRequest, LoginResponse>
+        implements AuthorizationRequestHandler {
+    public String getRouteKey(Request<?> msg) {
+        return null;
+    }
+
+    public String getName() {
+        return "login";
+    }
+
+    public Response<LoginResponse> doHandle(Request<?> requestObj) {
+        return null;
+    }
+}
