+++ softgroup/services/authorization/authorization-impl/src/main/java/handler/AuthorizationRequestRegisterHandler.java
@@ -1,0 +1,28 @@
+package handler;
+
+import com.softgroup.authorization.api.message.RegisterRequest;
+import com.softgroup.authorization.api.message.RegisterResponse;
+import com.softgroup.authorization.api.router.AuthorizationRequestHandler;
+import com.softgroup.authorization.api.router.AuthorizationRouterHandler;
+import com.softgroup.common.protocol.Request;
+import com.softgroup.common.protocol.Response;
+import com.softgroup.common.router.api.AbstractRequestHandler;
+import org.springframework.stereotype.Component;
+
+/**
+ * Created by polev on 25.02.2017.
+ */
+
+@Component
+public class AuthorizationRequestRegisterHandler
+        extends AbstractRequestHandler<RegisterRequest, RegisterResponse>
+        implements AuthorizationRequestHandler {
+
+    public String getName() {
+        return "register";
+    }
+
+    public Response<RegisterResponse> doHandle(Request<?> requestObj) {
+        return null;
+    }
+}
